FPLA-290 code review comments
