+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/casesubmission/CaseSubmissionTemplateDataGenerationServiceTest.java
@@ -387,6 +387,7 @@ public class CaseSubmissionTemplateDataGenerationServiceTest {
             CaseData updatedCaseData = givenCaseData.toBuilder()
                 .respondents1(wrapElements(Respondent.builder()
                     .party(RespondentParty.builder()
+                        .dateOfBirth(NOW.minusYears(34))
                         .address(Address.builder()
                             .postcode("SL11GF")
                             .build())
@@ -410,6 +411,7 @@ public class CaseSubmissionTemplateDataGenerationServiceTest {
             CaseData updatedCaseData = givenCaseData.toBuilder()
                 .respondents1(wrapElements(Respondent.builder()
                     .party(RespondentParty.builder()
+                        .dateOfBirth(NOW.minusYears(34))
                         .address(Address.builder()
                             .addressLine1("Flat 13")
                             .postcode("SL11GF")
